requirements.txt
astroid==2.3.1
certifi==2019.9.11
chardet==3.0.4
Django==2.1
djangorestframework==3.10.3
hashids==1.2.0
idna==2.8
isort==4.3.21
lazy-object-proxy==1.4.2
mccabe==0.6.1
pkg-resources==0.0.0
pylint==2.4.2
pylint-django==2.0.11
pylint-plugin-utils==0.6
pytz==2019.3
selenium==3.14.0
six==1.12.0
typed-ast==1.4.0
urllib3==1.25.6
wrapt==1.11.2
